Fix PR Review Comments
